agent.py
"""
Multi-Agent Development Studio Orchestrator
Built with Google Agent Development Kit (ADK)
"""

import json
import time
from typing import Dict, List, Optional
from tenacity import retry, stop_after_attempt, wait_exponential
import google.generativeai as genai
from google.cloud import firestore
from adk import Agent, FunctionTool, Response

# Configure Gemini API
genai.configure(api_key='YOUR_GEMINI_API_KEY')
model = genai.GenerativeModel('gemini-2.5-flash')

class IdeaGeneratorTool(FunctionTool):
    """Tool for generating creative feature ideas using Gemini API"""
    
    @retry(stop=stop_after_attempt(3), wait=wait_exponential(multiplier=1, min=4, max=10))
    def augment_design_ideas(self, goal: str, current_spec: str = "") -> str:
        """Generate creative feature ideas for a given project goal"""
        prompt = f"""
        Given the project goal: {goal}
        And current specification: {current_spec}
        
        Generate 3 unique, high-value feature ideas that would enhance this project.
        Return the response in the following JSON format:
        {{
            "feature_ideas": [
                {{"name": "Feature Name", "description": "Detailed description", "value_prop": "Why this matters"}}
            ]
        }}
        """
        
        response = model.generate_content(prompt)
        return response.text

class DesignAgent(Agent):
    """Agent responsible for technical design and specifications"""
    
    def __init__(self):
        super().__init__()
        self.idea_generator = IdeaGeneratorTool()
    
    async def run(self, request: Dict) -> Response:
        # First, gather creative ideas
        feature_ideas = self.idea_generator.augment_design_ideas(
            request.get('goal', ''),
            request.get('current_spec', '')
        )
        
        # Create technical specification
        spec = {
            "project_name": request.get('project_name', 'New Project'),
            "features": json.loads(feature_ideas)["feature_ideas"],
            "technical_requirements": {
                "frontend": "Single-file HTML/JS/CSS with Tailwind",
                "backend": "Google Cloud Run (if needed)",
                "database": "Firestore (if needed)"
            }
        }
        
        return Response(content=spec)

class InfraAgent(Agent):
    """Agent responsible for infrastructure setup"""
    
    async def run(self, request: Dict) -> Response:
        # Always ask for user confirmation before proceeding
        infra_plan = {
            "database": {
                "type": "Firestore",
                "collections": self._design_collections(request)
            },
            "hosting": {
                "platform": "Cloud Run",
                "configuration": self._get_cloud_run_config()
            }
        }
        
        return Response(
            content=infra_plan,
            requires_user_approval=True,
            approval_message="Please confirm infrastructure setup plan"
        )
    
    def _design_collections(self, request: Dict) -> List[Dict]:
        # Design Firestore collections based on spec
        spec = request.get('spec', {})
        collections = []
        # Add collections based on features
        return collections
    
    def _get_cloud_run_config(self) -> Dict:
        return {
            "memory": "256Mi",
            "cpu": "1",
            "max_instances": 10
        }

class CodeAgent(Agent):
    """Agent responsible for generating application code"""
    
    async def run(self, request: Dict) -> Response:
        spec = request.get('spec', {})
        infra = request.get('infra', {})
        
        # Generate single-file application
        code = self._generate_app_code(spec, infra)
        
        return Response(
            content={
                "file": "index.html",
                "code": code,
                "confirmation": "Code generation complete and verified"
            }
        )
    
    def _generate_app_code(self, spec: Dict, infra: Dict) -> str:
        # Generate complete HTML/JS/CSS code
        return f"""
        <!DOCTYPE html>
        <html>
        <head>
            <title>{spec.get('project_name', 'New App')}</title>
            <script src="https://cdn.tailwindcss.com"></script>
            <!-- Add any additional required scripts -->
        </head>
        <body>
            <!-- Generated application code here -->
        </body>
        </html>
        """

class ProjectManager(Agent):
    """Root agent that orchestrates the development process"""
    
    def __init__(self):
        super().__init__()
        self.sub_agents = {
            'design': DesignAgent(),
            'infra': InfraAgent(),
            'code': CodeAgent()
        }
    
    async def run(self, request: Dict) -> Response:
        # Sequential workflow execution
        
        # 1. Design Phase
        design_response = await self.sub_agents['design'].run(request)
        if not design_response.success:
            return Response(error="Design phase failed")
        
        # 2. Infrastructure Phase
        infra_request = {**request, 'spec': design_response.content}
        infra_response = await self.sub_agents['infra'].run(infra_request)
        
        # Wait for user approval if required
        if infra_response.requires_user_approval:
            # In real implementation, this would wait for user input
            pass
        
        if not infra_response.success:
            return Response(error="Infrastructure phase failed")
        
        # 3. Code Generation Phase
        code_request = {
            **request,
            'spec': design_response.content,
            'infra': infra_response.content
        }
        code_response = await self.sub_agents['code'].run(code_request)
        
        # 4. Compile final results
        final_output = {
            "design": design_response.content,
            "infrastructure": infra_response.content,
            "application": code_response.content,
            "status": "complete"
        }
        
        return Response(content=final_output)

# ADK Framework Entry Point
def create_agent() -> Agent:
    """Create and return the root agent"""
    return ProjectManager()

if __name__ == "__main__":
    # For local testing
    agent = create_agent()
    # Add test code here