requirements.txt
adk==0.0.5
google-cloud-firestore==2.13.1
tenacity==8.2.3
google-generativeai==0.2.2
google-ai-generativelanguage==0.3.3
protobuf>=3.19.5,<4.0.0dev
python-dotenv==1.0.0